Create imageShell.sh
This script will inject an image with a PHP shell that you can interact with to upload a reverse shell. It utilizes the exiftool to pass the PHP one-liner shell into the "comment" section of the image of choice.

I made this for educational purpose and to speedup the image to shell process.